[rename] #20 gifter 도메인 이름을 room으로 변경
