--- src/main/java/org/sopt/sweet/domain/gifter/repository/GifterRepository.java
@@ -1,7 +1,0 @@
-package org.sopt.sweet.domain.gifter.repository;
-
-import org.sopt.sweet.domain.gifter.entity.Gifter;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface GifterRepository extends JpaRepository<Gifter, Long> {
-}
