+++ src/main/java/org/sopt/sweet/domain/room/controller/RoomApi.java
@@ -1,4 +1,4 @@
-package org.sopt.sweet.domain.gifter.controller;
+package org.sopt.sweet.domain.room.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -7,12 +7,12 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import org.sopt.sweet.domain.gifter.dto.request.PresignedURLRequestDto;
+import org.sopt.sweet.domain.room.dto.request.PresignedURLRequestDto;
 import org.sopt.sweet.global.common.SuccessResponse;
 import org.springframework.http.ResponseEntity;
 
-@Tag(name = "선물 모임", description = "선물 모임 관련 API")
-public interface GifterApi {
+@Tag(name = "선물방", description = "선물방 관련 API")
+public interface RoomApi {
 
     @ApiResponses(
             value = {
