+++ src/main/java/org/sopt/sweet/domain/room/controller/RoomController.java
@@ -1,9 +1,9 @@
-package org.sopt.sweet.domain.gifter.controller;
+package org.sopt.sweet.domain.room.controller;
 
 import lombok.RequiredArgsConstructor;
-import org.sopt.sweet.domain.gifter.dto.request.PresignedURLRequestDto;
-import org.sopt.sweet.domain.gifter.dto.response.PresignedURLResponseDto;
-import org.sopt.sweet.domain.gifter.service.GifterService;
+import org.sopt.sweet.domain.room.dto.request.PresignedURLRequestDto;
+import org.sopt.sweet.domain.room.dto.response.PresignedURLResponseDto;
+import org.sopt.sweet.domain.room.service.RoomService;
 import org.sopt.sweet.global.common.SuccessResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -12,15 +12,15 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RequiredArgsConstructor
-@RequestMapping("/api/gifter")
+@RequestMapping("/api/room")
 @RestController
-public class GifterController implements GifterApi {
+public class RoomController implements RoomApi {
 
-    private final GifterService gifterService;
+    private final RoomService roomService;
 
     @PostMapping("/presignedURL")
     public ResponseEntity<SuccessResponse<?>> getPresignedURL(@RequestBody PresignedURLRequestDto presignedURLRequestDto) {
-        final PresignedURLResponseDto presignedURLResponseDto = gifterService.getPresignedURL(presignedURLRequestDto);
+        final PresignedURLResponseDto presignedURLResponseDto = roomService.getPresignedURL(presignedURLRequestDto);
         return SuccessResponse.created(presignedURLResponseDto);
     }
 }
