+++ src/main/java/org/sopt/sweet/domain/room/entity/RoomMember.java
@@ -1,6 +1,5 @@
-package org.sopt.sweet.domain.gifter.entity;
+package org.sopt.sweet.domain.room.entity;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
 import lombok.*;
 import org.sopt.sweet.domain.member.entity.Member;
@@ -10,16 +9,16 @@ import org.sopt.sweet.domain.member.entity.Member;
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @Builder(access = AccessLevel.PRIVATE)
 @Entity
-public class GifterMember {
+public class RoomMember {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "gifter_member_id")
+    @Column(name = "room_member_id")
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "gifter_id")
-    private Gifter gifter;
+    @JoinColumn(name = "room_id")
+    private Room room;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id", nullable = false)
