+++ src/main/java/org/sopt/sweet/domain/room/service/RoomService.java
@@ -1,8 +1,8 @@
-package org.sopt.sweet.domain.gifter.service;
+package org.sopt.sweet.domain.room.service;
 
 import lombok.RequiredArgsConstructor;
-import org.sopt.sweet.domain.gifter.dto.request.PresignedURLRequestDto;
-import org.sopt.sweet.domain.gifter.dto.response.PresignedURLResponseDto;
+import org.sopt.sweet.domain.room.dto.request.PresignedURLRequestDto;
+import org.sopt.sweet.domain.room.dto.response.PresignedURLResponseDto;
 import org.sopt.sweet.global.config.FileService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -10,7 +10,7 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @RequiredArgsConstructor
 @Transactional
-public class GifterService {
+public class RoomService {
 
     private final FileService fileService;
 
@@ -18,4 +18,5 @@ public class GifterService {
         String URL = fileService.getPreSignedUrl("gifterImg", presignedURLRequestDto.fileName());
         return PresignedURLResponseDto.of(URL);
     }
+
 }
